requirements.txt
opencv_python==4.5.5.62
PyYAML==6.0
Telethon==1.24.0
